requirements.txt
torch
torchvision
datasets
Pillow
tqdm